refactor(data-point/helpers): Rename type-check-function-reducers.js to type-check-functions.js
