+++ packages/data-point/lib/helpers/type-check-helpers.js
@@ -1,4 +1,4 @@
-const typeCheckFunctions = require('./type-check-function-reducers')
+const typeCheckFunctions = require('./type-check-functions')
 
 const modifiers = {
   string: typeCheckFunctions.isString,
